dom/base/test/test_bug416383.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=416383
-->
<head>
  <title>Test for Bug 416383</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=416383">Mozilla Bug 416383</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<div id="test_parent"><div someattr="foo"></div></div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 416383 **/

function runTest() {
  var testParent = document.getElementById("test_parent");
  var attrs = testParent.firstChild.attributes;
  ok(attrs != null, "Element should have attributes!");
  var attr = testParent.firstChild.getAttributeNode("someattr");
  ok(attr.value == "foo", "The value of the attribute should be 'foo'!");
  testParent.firstChild.remove();
  SpecialPowers.gc();
  ok(true, "Browser should not crash!")

}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
